quantum-autotrading/requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 비동기 처리
asyncio-mqtt==0.16.1
websockets==12.0
aiohttp==3.9.1

# 데이터 처리 및 분석
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# 기술적 지표 계산
ta-lib==0.4.28
talib-binary==0.4.19

# 데이터베이스
supabase==2.3.0
asyncpg==0.29.0
redis==5.0.1

# HTTP 클라이언트
httpx==0.25.2
requests==2.31.0

# 환경 설정
python-dotenv==1.0.0

# 로깅
loguru==0.7.2

# 시간 처리
python-dateutil==2.8.2
pytz==2023.3

# 암호화 및 보안
cryptography==41.0.8
pyjwt==2.8.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# 모니터링
prometheus-client==0.19.0

# 기타 유틸리티
python-multipart==0.0.6
email-validator==2.1.0